+++ src/main/java/com/efonemax/widgetREST/repository/WidgetRepository.java
@@ -3,5 +3,5 @@ package com.efonemax.widgetREST.repository;
 import com.efonemax.widgetREST.domain.Widget;
 import org.springframework.data.repository.CrudRepository;
 
-public interface WidgetRepository extends CrudRepository<Widget, Long> {
+public interface WidgetRepository extends CrudRepository<Widget, Integer> {
 }
